Move iOS and macOS implementations into new `apple` module
