+++ src/platform_impl/apple/appkit/app_state.rs
