+++ src/platform_impl/apple/appkit/monitor.rs
