+++ src/platform_impl/apple/appkit/observer.rs
