+++ src/platform_impl/apple/appkit/window_delegate.rs
