+++ src/platform_impl/apple/uikit/app_delegate.rs
