+++ src/platform_impl/apple/uikit/app_state.rs
